Update README.md
    Problem

1.1 Description

In this homework, you will have the complete reinforcement-learning experience: training an agent from scratch to solve a simple domain using Q-learning.

The environment you will be applying Q-learning to is called Taxi (Taxi-v3). The Taxi problem was intro-duced by Dietterich 1998 and has been used for reinforcement-learning research in the past. It is a grid-based environment where the goal of the agent is to pick up a passenger at one location and drop them o at another.

The map is xed and the environment has deterministic transitions. However, the distinct pickup and drop-o points are chosen randomly from 4 xed locations in the grid, each assigned a di erent letter. The starting location of the taxicab is also chosen randomly.

The agent has 6 actions: 4 for movement, 1 for pickup, and 1 for drop-o . Attempting a pickup when there is no passenger at the location incurs a reward of -10. Dropping o a passenger outside one of the four designated zones is prohibited, and attempting it also incurs a reward of 10. Dropping the passenger o at the correct destination provides the agent with a reward of 20. Otherwise, the agent incurs a reward of 1 per time step.

Your job is to train your agent until it converges to the optimal state-action value function. You will have to think carefully about algorithm implementation, especially exploration parameters.